+++ src/gun.cpp
@@ -30,8 +30,8 @@ void Shot::makeShot(Space_Object pos) {
 }
 int Gun::gun_manager(Space_Object pos, int signal, int mode) {
   if (shots.size() > 10) signal = 0;
-  for (int i = 0; i < shots.size(); i++) eraseShots(shots[i]);
-  for (int i = 0; i < shots.size(); i++) moveShotsRigth(this, shots[i], i);
+  for (long unsigned int i = 0; i < shots.size(); i++) eraseShots(shots[i]);
+  for (long unsigned int i = 0; i < shots.size(); i++) moveShotsRigth(this, shots[i], i);
   char shot = {'-'};
   Space_Object shotpos(pos.getX(), pos.getY());
   if (signal) {
@@ -40,6 +40,6 @@ int Gun::gun_manager(Space_Object pos, int signal, int mode) {
       while (shots.size() < 5) shots.push_back(new Shot(shot, shotpos));
     }
   }
-  for (int i = 0; i < shots.size(); i++) drawShots(shots[i]);
+  for (long unsigned int i = 0; i < shots.size(); i++) drawShots(shots[i]);
   return signal;
 }
